+++ doc/source/conf.py
@@ -16,11 +16,16 @@ author = "Jisoo Song"
 extensions = [
     "sphinx.ext.autodoc",
     "sphinx.ext.viewcode",
+    "sphinx.ext.intersphinx",
     "numpydoc",
     "matplotlib.sphinxext.plot_directive",
 ]
 
 autodoc_member_order = "bysource"
+
+intersphinx_mapping = {
+    "scipy": ("https://docs.scipy.org/doc/scipy/", None),
+}
 
 numpydoc_use_plots = True
 numpydoc_show_class_members = False