+++ src/curvesimilarities/__init__.py
@@ -3,6 +3,7 @@
 from .averagefrechet import afd, afd_owp, qafd, qafd_owp
 from .dtw import dtw, dtw_owp, sdtw, sdtw_owp
 from .frechet import dfd, fd
+from .hausdorff import hd
 from .integfrechet import ifd, ifd_owp
 
 __all__ = [
@@ -18,4 +19,5 @@ __all__ = [
     "afd_owp",
     "qafd",
     "qafd_owp",
+    "hd",
 ]