+++ src/curvesimilarities/hausdorff.py
@@ -1,0 +1,15 @@
+"""Hausdorff distance."""
+
+__all__ = [
+    "hd",
+]
+
+
+def hd(P, Q):
+    """(Continuous) Hausdorff distance between two open polygonal curves.
+
+    See Also
+    --------
+    scipy.spatial.distance.directed_hausdorff :
+        Directed Hausdorff distance between discrete sets of points.
+    """